[Fix] middleware refactoring
