+++ bin/www
@@ -13,7 +13,7 @@ require("dotenv").config();
  * Get port from environment and store in Express.
  */
 
-const port = normalizePort(PORT || "3001");
+const port = normalizePort(PORT || "3000");
 app.set("port", port);
 
 /**
