+++ common/envConstants.js
@@ -2,7 +2,7 @@ const dotenv = require("dotenv");
 dotenv.config();
 
 const constants = {
-  PORT: process.env.PORT,
+  PORT: process.env.PORT || "3000",
   DB_HOST: process.env.DB_HOST,
   DB_USER: process.env.DB_USER,
   DB_PORT: process.env.DB_PORT,
