+++ routes/saveTestResultRouter.js
@@ -2,7 +2,7 @@ const express = require("express");
 const {
   saveTestResultPage,
 } = require("../controllers/saveTestResultController");
-const authMiddleware = require("../middlewares/authMiddleware");
+const authMiddleware = require("../middleware/authMiddleware");
 
 const router = express.Router();
 router.put("/", authMiddleware, saveTestResultPage);
